requirements.txt
langchain
langchain-core
langchain-community
openai
ollama
python-dotenv
docx